requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.2.1
pyspark==3.5.4
sqlparse==0.5.3
